chore: move into success.spec.js
